pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "freqinout"
version = "1.0.0"
description = "HF Radio Frequency and Net Control Utility"
readme = "README.md"
authors = [{ name = "Bill Daniel" }]
license = { text = "MIT OR GPL-3.0-or-later" }
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "pytz>=2023.3",
  "psutil>=5.9.0",
  "PySide6>=6.7.0",
  "PyYAML>=6.0",
]

[project.scripts]
freqinout = "freqinout.main:main"

[tool.setuptools]
packages = ["freqinout"]
